Attempt to make Facebook and Google logins do so with minimal use of Browser launching.
In the case of facebook, it can use the APP.  I don't know what fixed it but when I made a new app in developers.facebook.com/apps and then changed the keys in AWSKeys.xcconfig and updated the app-Id for facebook in the cognito federated identity pool, it just started working.  For some reason prior to that it did not list the app in the facebook apps that could be authorized.

In the case of Google, it always asked for confirmation on login, now it uses saved keychain authentication information if it exists.